requirements.txt
matrix-nio>=0.24.0
python-dotenv>=1.0.0
tomli>=2.0.1
pytest-asyncio>=0.23.0
pytest>=7.0.0
anthropic>=0.25.0
aiohttp>=3.9.0
aiofiles>=23.2.0
PyYAML>=6.0.0
psutil>=5.9.0
# Optional (uncomment to install) for faster event loop on mac/linux
uvloop>=0.19.0